+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add macro git:lastUpdated
+
 ## [1.0.1] 2019-09-13
 
 - Locate files with require.resolve()